srcs/services/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.6
django-cors-headers==4.2.0  # or any version >=4.2.0 that supports Python 3.9
gunicorn==23.0.0